Move RuntimeType back to shared Enum
